Identified a bug

+++ Assets/Scripts/PooledObject.cs
@@ -2,6 +2,6 @@ using UnityEngine;
 
 public class PooledObject : MonoBehaviour
 {
-    public short PoolIndex;
+    public int PoolIndex;
 }
